+++ test/helpers.test.js
@@ -11,5 +11,18 @@ describe('timestampToBuffer()', () => {
     expect(intToBuffer(31).toString('hex')).to.eq('1f')
     expect(intToBuffer(256).toString('hex')).to.eq('0100')
     expect(intToBuffer(1604499020).toString('hex')).to.eq('5fa2b64c')
+    expect(intToBuffer('1604499020').toString('hex')).to.eq('5fa2b64c')
+  })
+})
+
+describe('Buffer must be align with other languages SDKs', () => {
+  const stringsToHex = arr => Buffer.concat(arr.map(Buffer.from)).toString('hex')
+
+  it('expect have compatible result', async () => {
+    expect(stringsToHex([''])).to.eq('')
+    expect(stringsToHex(['1'])).to.eq('31')
+    expect(stringsToHex(['1', '2'])).to.eq('3132')
+    expect(stringsToHex(['', '1'])).to.eq('31')
+    expect(stringsToHex(['żźćłóęąń', '0'])).to.eq('c5bcc5bac487c582c3b3c499c485c58430')
   })
 })